==> 7_Combine_and_Split_Text_Data.ipynb <==
## Structured and text data
 
- Already separate datasets for outcomes split by training, test 
- Combine structured data with text features and list of strings
- Aim: create final dataset to feed into model
- NB NOT including tfidf and topic modelling - these are completed as transformer steps in the gridsearchcv pipeline in the analysis notebooks - as otherwise there would be data leakage between cross validation folds


# Load user-written functions

get_ipython().run_line_magic('load_ext', 'autoreload')
get_ipython().run_line_magic('autoreload', '2')

import analysis_functions

# Bring in text features (vulnerabilities, concreteness, emotions NOT tfidf or topics)
import pickle
import os

# Set working directory
os.chdir("C:\\Users\\[username]\\Downloads\\Updated Text Data") # insert [username]

filename = open("Created/text_rq1_vulnerabilities.pkl", "rb")
df_text_features_vulnerabilities_rq1 = pickle.load(filename)
print(df_text_features_vulnerabilities_rq1.shape)
print(df_text_features_vulnerabilities_rq1.index)


filename = open("Created/text_rq2_vulnerabilities.pkl", "rb")
df_text_features_vulnerabilities_rq2 = pickle.load(filename)
print(df_text_features_vulnerabilities_rq2.shape)
print(df_text_features_vulnerabilities_rq2.index)

import pickle
import os

# Set working directory

# Import structured data (with all the final data cleaning but before it's split)
# These datasets created in 5a_Combine_Data
filename = open("../Updated Structured Data/Created/df_outcome1_before_splitting.pkl", "rb")
df_outcome1_before_splitting = pickle.load(filename)
print(df_outcome1_before_splitting.shape)


filename = open("../Updated Structured Data/Created/df_outcome2_before_splitting.pkl", "rb")
df_outcome2_before_splitting = pickle.load(filename)
print(df_outcome2_before_splitting.shape)


# Select the same rows as the cleaned structured data
# By merging on the key cols
import pandas as pd
# Format 
df_text_features_vulnerabilities_rq1['ReferralDatetime'] = pd.to_datetime(df_text_features_vulnerabilities_rq1['ReferralDatetime'])
df_text_features_vulnerabilities_rq2['ReferralDatetime'] = pd.to_datetime(df_text_features_vulnerabilities_rq2['ReferralDatetime'])


# rq1 
df_text_features_vulnerabilities_rq1 = pd.merge(df_outcome1_before_splitting[['PSID', 'ReferralDatetime']], df_text_features_vulnerabilities_rq1, on = ['PSID', 'ReferralDatetime'], how = 'left')
df_text_features_vulnerabilities_rq1.drop_duplicates(subset = ['PSID', 'ReferralDatetime'], inplace = True)
df_text_features_vulnerabilities_rq1.sort_values(by = 'ReferralDatetime', inplace = True)
print(df_text_features_vulnerabilities_rq1.shape)


# rq2
df_text_features_vulnerabilities_rq2 = pd.merge(df_outcome2_before_splitting[['PSID', 'ReferralDatetime']], df_text_features_vulnerabilities_rq2, on = ['PSID', 'ReferralDatetime'], how = 'left')
df_text_features_vulnerabilities_rq2.drop_duplicates(subset = ['PSID', 'ReferralDatetime'], inplace = True)
df_text_features_vulnerabilities_rq2.sort_values(by = 'ReferralDatetime', inplace = True)
print(df_text_features_vulnerabilities_rq2.shape)

import pickle

filename = open("../Data for Model/X_train_rq1_ts_str.pkl", "rb")
X_train_rq1_ts = pickle.load(filename)
print(X_train_rq1_ts.shape)
print(X_train_rq1_ts.index)

filename = open("../Data for Model/X_test_rq1_ts_str.pkl", "rb")
X_test_rq1_ts = pickle.load(filename)
print(X_test_rq1_ts.shape)
print(X_test_rq1_ts.index)

filename = open("../Data for Model/X_train_rq1_ss_str.pkl", "rb")
X_train_rq1_ss = pickle.load(filename)
print(X_train_rq1_ss.shape)
print(X_train_rq1_ss.index)

filename = open("../Data for Model/X_test_rq1_ss_str.pkl", "rb")
X_test_rq1_ss = pickle.load(filename)
print(X_test_rq1_ss.shape)
print(X_test_rq1_ss.index)


import pickle

filename = open("../Data for Model/X_train_rq2_ts_str.pkl", "rb")
X_train_rq2_ts = pickle.load(filename)
print(X_train_rq2_ts.shape)
print(X_train_rq2_ts.index)

filename = open("../Data for Model/X_test_rq2_ts_str.pkl", "rb")
X_test_rq2_ts = pickle.load(filename)
print(X_test_rq2_ts.shape)
print(X_test_rq2_ts.index)

filename = open("../Data for Model/X_train_rq2_ss_str.pkl", "rb")
X_train_rq2_ss = pickle.load(filename)
print(X_train_rq2_ss.shape)
print(X_train_rq2_ss.index)

filename = open("../Data for Model/X_test_rq2_ss_str.pkl", "rb")
X_test_rq2_ss = pickle.load(filename)
print(X_test_rq2_ss.shape)
print(X_test_rq2_ss.index)


# Merge df_text_features_vulnerabilities with train, test,  
# Merge on left with the train etc data on the left so just those rows picked up

df_text_features_vulnerabilities_rq1_ts_train = pd.merge(X_train_rq1_ts, df_text_features_vulnerabilities_rq1, how = 'left', on = ['PSID', 'ReferralDatetime'])
df_text_features_vulnerabilities_rq1_ts_train.reset_index(inplace = True, drop = True)
print(df_text_features_vulnerabilities_rq1_ts_train.shape)
print(X_train_rq1_ts.shape)
print(df_text_features_vulnerabilities_rq1_ts_train.index)

df_text_features_vulnerabilities_rq1_ts_test = pd.merge(X_test_rq1_ts, df_text_features_vulnerabilities_rq1, how = 'left', on = ['PSID', 'ReferralDatetime'])
df_text_features_vulnerabilities_rq1_ts_test.reset_index(inplace = True, drop = True)
print(df_text_features_vulnerabilities_rq1_ts_test.shape)
print(X_test_rq1_ts.shape)
print(df_text_features_vulnerabilities_rq1_ts_test.index)

df_text_features_vulnerabilities_rq1_ss_train = pd.merge(X_train_rq1_ss, df_text_features_vulnerabilities_rq1, how = 'left', on = ['PSID', 'ReferralDatetime'])
df_text_features_vulnerabilities_rq1_ss_train.reset_index(inplace = True, drop = True)
print(df_text_features_vulnerabilities_rq1_ss_train.shape)
print(X_train_rq1_ss.shape)
print(df_text_features_vulnerabilities_rq1_ss_train.index)

df_text_features_vulnerabilities_rq1_ss_test = pd.merge(X_test_rq1_ss, df_text_features_vulnerabilities_rq1, how = 'left', on = ['PSID', 'ReferralDatetime'])
df_text_features_vulnerabilities_rq1_ss_test.reset_index(inplace = True, drop = True)
print(df_text_features_vulnerabilities_rq1_ss_test.shape)
print(X_test_rq1_ss.shape)
print(df_text_features_vulnerabilities_rq1_ss_test.index)


# Save training, test  data

# RQ1
with open("../Data for Model/df_text__vulnerabilities_rq1_ts_train.pkl", "wb") as handle:
    pickle.dump(df_text_features_vulnerabilities_rq1_ts_train, handle, protocol = pickle.HIGHEST_PROTOCOL)
    
with open("../Data for Model/df_text_features_vulnerabilities_rq1_ts_test.pkl", "wb") as handle:
    pickle.dump(df_text_features_vulnerabilities_rq1_ts_test, handle, protocol = pickle.HIGHEST_PROTOCOL)

    
with open("../Data for Model/df_text_features_vulnerabilities_rq1_ss_train.pkl", "wb") as handle:
    pickle.dump(df_text_features_vulnerabilities_rq1_ss_train, handle, protocol = pickle.HIGHEST_PROTOCOL)
    
with open("../Data for Model/df_text_features_vulnerabilities_rq1_ss_test.pkl", "wb") as handle:
    pickle.dump(df_text_features_vulnerabilities_rq1_ss_test, handle, protocol = pickle.HIGHEST_PROTOCOL)


# Merge df_text_features_vulnerabilities with train, test,  
# Merge on left with the train etc data on the left so just those rows picked up

df_text_features_vulnerabilities_rq2_ts_train = pd.merge(X_train_rq2_ts, df_text_features_vulnerabilities_rq2, how = 'left', on = ['PSID', 'ReferralDatetime'])
df_text_features_vulnerabilities_rq2_ts_train.reset_index(inplace = True, drop = True)
print(df_text_features_vulnerabilities_rq2_ts_train.shape)
print(X_train_rq2_ts.shape)
print(df_text_features_vulnerabilities_rq2_ts_train.index)

df_text_features_vulnerabilities_rq2_ts_test = pd.merge(X_test_rq2_ts, df_text_features_vulnerabilities_rq2, how = 'left', on = ['PSID', 'ReferralDatetime'])
df_text_features_vulnerabilities_rq2_ts_test.reset_index(inplace = True, drop = True)
print(df_text_features_vulnerabilities_rq2_ts_test.shape)
print(X_test_rq2_ts.shape)
print(df_text_features_vulnerabilities_rq2_ts_test.index)

df_text_features_vulnerabilities_rq2_ss_train = pd.merge(X_train_rq2_ss, df_text_features_vulnerabilities_rq2, how = 'left', on = ['PSID', 'ReferralDatetime'])
df_text_features_vulnerabilities_rq2_ss_train.reset_index(inplace = True, drop = True)
print(df_text_features_vulnerabilities_rq2_ss_train.shape)
print(X_train_rq2_ss.shape)
print(df_text_features_vulnerabilities_rq2_ss_train.index)

df_text_features_vulnerabilities_rq2_ss_test = pd.merge(X_test_rq2_ss, df_text_features_vulnerabilities_rq2, how = 'left', on = ['PSID', 'ReferralDatetime'])
df_text_features_vulnerabilities_rq2_ss_test.reset_index(inplace = True, drop = True)
print(df_text_features_vulnerabilities_rq2_ss_test.shape)
print(X_test_rq2_ss.shape)
print(df_text_features_vulnerabilities_rq2_ss_test.index)



# Save training, test  data

# rq2
with open("../Data for Model/df_text_features_vulnerabilities_rq2_ts_train.pkl", "wb") as handle:
    pickle.dump(df_text_features_vulnerabilities_rq2_ts_train, handle, protocol = pickle.HIGHEST_PROTOCOL)
    
with open("../Data for Model/df_text_features_vulnerabilities_rq2_ts_test.pkl", "wb") as handle:
    pickle.dump(df_text_features_vulnerabilities_rq2_ts_test, handle, protocol = pickle.HIGHEST_PROTOCOL)
    
with open("../Data for Model/df_text_features_vulnerabilities_rq2_ss_train.pkl", "wb") as handle:
    pickle.dump(df_text_features_vulnerabilities_rq2_ss_train, handle, protocol = pickle.HIGHEST_PROTOCOL)
    
with open("../Data for Model/df_text_features_vulnerabilities_rq2_ss_test.pkl", "wb") as handle:
    pickle.dump(df_text_features_vulnerabilities_rq2_ss_test, handle, protocol = pickle.HIGHEST_PROTOCOL)


import pickle
import os

# Set working directory
os.chdir("C:\\Users\\[username]\\Downloads\\Updated Text Data") # insert [username]

filename = open("../Updated Structured Data/Created/df_tfidf_topics_train_rq1_ts_all.pkl", "rb")
df_tfidf_topics_rq1_ts_train = pickle.load(filename)
# Referral_Start_Date dropped to avoid duplicate (don't match on it with so many missing)
#df_tfidf_topics_rq1_ts_train.drop(columns = 'ReferralDatetime', inplace = True)
print(df_tfidf_topics_rq1_ts_train.shape)
print(df_tfidf_topics_rq1_ts_train.index)

filename = open("../Updated Structured Data/Created/df_tfidf_topics_test_rq1_ts_all.pkl", "rb")
df_tfidf_topics_rq1_ts_test = pickle.load(filename)
#df_tfidf_topics_rq1_ts_test.drop(columns = 'ReferralDatetime', inplace = True)
print(df_tfidf_topics_rq1_ts_test.shape)
print(df_tfidf_topics_rq1_ts_test.index)

filename = open("../Updated Structured Data/Created/df_tfidf_topics_train_rq1_ss_all.pkl", "rb")
df_tfidf_topics_rq1_ss_train = pickle.load(filename)
#df_tfidf_topics_rq1_ss_train.drop(columns = 'ReferralDatetime', inplace = True)
print(df_tfidf_topics_rq1_ss_train.shape)
print(df_tfidf_topics_rq1_ss_train.index)

filename = open("../Updated Structured Data/Created/df_tfidf_topics_test_rq1_ss_all.pkl", "rb")
df_tfidf_topics_rq1_ss_test = pickle.load(filename)
#df_tfidf_topics_rq1_ss_test.drop(columns = 'ReferralDatetime', inplace = True)
print(df_tfidf_topics_rq1_ss_test.shape)
print(df_tfidf_topics_rq1_ss_test.index)

import pickle
import os

# Set working directory
os.chdir("C:\\Users\\[username]\\Downloads\\Updated Text Data") # insert [username]

filename = open("../Updated Structured Data/Created/df_tfidf_topics_train_rq2_ts_all.pkl", "rb")
df_tfidf_topics_rq2_ts_train = pickle.load(filename)
# Referral_Start_Date dropped to avoid duplicate (don't match on it with so many missing)
#df_tfidf_topics_rq2_ts_train.drop(columns = 'ReferralDatetime', inplace = True)
print(df_tfidf_topics_rq2_ts_train.shape)
print(df_tfidf_topics_rq2_ts_train.index)

filename = open("../Updated Structured Data/Created/df_tfidf_topics_test_rq2_ts_all.pkl", "rb")
df_tfidf_topics_rq2_ts_test = pickle.load(filename)
#df_tfidf_topics_rq2_ts_test.drop(columns = 'ReferralDatetime', inplace = True)
print(df_tfidf_topics_rq2_ts_test.shape)
print(df_tfidf_topics_rq2_ts_test.index)

filename = open("../Updated Structured Data/Created/df_tfidf_topics_train_rq2_ss_all.pkl", "rb")
df_tfidf_topics_rq2_ss_train = pickle.load(filename)
#df_tfidf_topics_rq2_ss_train.drop(columns = 'ReferralDatetime', inplace = True)
print(df_tfidf_topics_rq2_ss_train.shape)
print(df_tfidf_topics_rq2_ss_train.index)

filename = open("../Updated Structured Data/Created/df_tfidf_topics_test_rq2_ss_all.pkl", "rb")
df_tfidf_topics_rq2_ss_test = pickle.load(filename)
#df_tfidf_topics_rq2_ss_test.drop(columns = 'ReferralDatetime', inplace = True)
print(df_tfidf_topics_rq2_ss_test.shape)
print(df_tfidf_topics_rq2_ss_test.index)

# Merge structured and text feature data with list of strings data
print(df_text_features_vulnerabilities_rq1_ts_train.shape)
print(df_tfidf_topics_rq1_ts_train.shape)
df_str_text_rq1_ts_train = pd.merge(df_text_features_vulnerabilities_rq1_ts_train, df_tfidf_topics_rq1_ts_train, how = 'left', on = ['PSID', 'ReferralDatetime'])
print(df_str_text_rq1_ts_train.shape)

print(df_text_features_vulnerabilities_rq1_ts_test.shape)
print(df_tfidf_topics_rq1_ts_test.shape)
df_str_text_rq1_ts_test = pd.merge(df_text_features_vulnerabilities_rq1_ts_test, df_tfidf_topics_rq1_ts_test, how = 'left', on = ['PSID', 'ReferralDatetime'])
print(df_str_text_rq1_ts_test.shape) 

print(df_text_features_vulnerabilities_rq1_ss_train.shape)
print(df_tfidf_topics_rq1_ss_train.shape)
df_str_text_rq1_ss_train = pd.merge(df_text_features_vulnerabilities_rq1_ss_train, df_tfidf_topics_rq1_ss_train, how = 'left', on = ['PSID', 'ReferralDatetime'])
print(df_str_text_rq1_ss_train.shape)


print(df_text_features_vulnerabilities_rq1_ss_test.shape)
print(df_tfidf_topics_rq1_ss_test.shape)
df_str_text_rq1_ss_test = pd.merge(df_text_features_vulnerabilities_rq1_ss_test, df_tfidf_topics_rq1_ss_test, how = 'left', on = ['PSID', 'ReferralDatetime'])
print(df_str_text_rq1_ss_test.shape)  


# Merge structured and text feature data with list of strings data

print(df_text_features_vulnerabilities_rq2_ts_train.shape)
print(df_tfidf_topics_rq2_ts_train.shape)
df_str_text_rq2_ts_train = pd.merge(df_text_features_vulnerabilities_rq2_ts_train, df_tfidf_topics_rq2_ts_train, how = 'left', on = ['PSID', 'ReferralDatetime'])
print(df_str_text_rq2_ts_train.shape)

print(df_text_features_vulnerabilities_rq2_ts_test.shape)
print(df_tfidf_topics_rq2_ts_test.shape)
df_str_text_rq2_ts_test = pd.merge(df_text_features_vulnerabilities_rq2_ts_test, df_tfidf_topics_rq2_ts_test, how = 'left', on = ['PSID', 'ReferralDatetime'])
print(df_str_text_rq2_ts_test.shape)    

print(df_text_features_vulnerabilities_rq2_ss_train.shape)
print(df_tfidf_topics_rq2_ss_train.shape)
df_str_text_rq2_ss_train = pd.merge(df_text_features_vulnerabilities_rq2_ss_train, df_tfidf_topics_rq2_ss_train, how = 'left', on = ['PSID', 'ReferralDatetime'])
print(df_str_text_rq2_ss_train.shape)

print(df_text_features_vulnerabilities_rq2_ss_test.shape)
print(df_tfidf_topics_rq2_ss_test.shape)
df_str_text_rq2_ss_test = pd.merge(df_text_features_vulnerabilities_rq2_ss_test, df_tfidf_topics_rq2_ss_test, how = 'left', on = ['PSID', 'ReferralDatetime'])
print(df_str_text_rq2_ss_test.shape)  
 

# Resetting indices in preparation for splits (splitters expect range index)
print(df_str_text_rq1_ts_train.index)
df_str_text_rq1_ts_train.reset_index(inplace = True, drop = True)
print(df_str_text_rq1_ts_train.index)
print(df_str_text_rq1_ts_test.index)
df_str_text_rq1_ts_test.reset_index(inplace = True, drop = True)
print(df_str_text_rq1_ts_test.index)

print(df_str_text_rq1_ss_train.index)
df_str_text_rq1_ss_train.reset_index(inplace = True, drop = True)
print(df_str_text_rq1_ss_train.index)
print(df_str_text_rq1_ss_test.index)
df_str_text_rq1_ss_test.reset_index(inplace = True, drop = True)
print(df_str_text_rq1_ss_test.index)



# Resetting indices in preparation for splits (splitters expect range index)
print(df_str_text_rq2_ts_train.index)
df_str_text_rq2_ts_train.reset_index(inplace = True, drop = True)
print(df_str_text_rq2_ts_train.index)
print(df_str_text_rq2_ts_test.index)
df_str_text_rq2_ts_test.reset_index(inplace = True, drop = True)
print(df_str_text_rq2_ts_test.index)

print(df_str_text_rq2_ss_train.index)
df_str_text_rq2_ss_train.reset_index(inplace = True, drop = True)
print(df_str_text_rq2_ss_train.index)
print(df_str_text_rq2_ss_test.index)
df_str_text_rq2_ss_test.reset_index(inplace = True, drop = True)
print(df_str_text_rq2_ss_test.index)


### Save all

# Save training, test  data

# RQ1

with open("../Data for Model/X_train_rq1_ts_all.pkl", "wb") as handle:
    pickle.dump(df_str_text_rq1_ts_train, handle, protocol = pickle.HIGHEST_PROTOCOL)
    
with open("../Data for Model/X_test_rq1_ts_all.pkl", "wb") as handle:
    pickle.dump(df_str_text_rq1_ts_test, handle, protocol = pickle.HIGHEST_PROTOCOL)
    
with open("../Data for Model/X_train_rq1_ss_all.pkl", "wb") as handle:
    pickle.dump(df_str_text_rq1_ss_train, handle, protocol = pickle.HIGHEST_PROTOCOL)
    
with open("../Data for Model/X_test_rq1_ss_all.pkl", "wb") as handle:
    pickle.dump(df_str_text_rq1_ss_test, handle, protocol = pickle.HIGHEST_PROTOCOL)


# Save training, test  data

# RQ2
with open("../Data for Model/X_train_rq2_ts_all.pkl", "wb") as handle:
    pickle.dump(df_str_text_rq2_ts_train, handle, protocol = pickle.HIGHEST_PROTOCOL)
    
with open("../Data for Model/X_test_rq2_ts_all.pkl", "wb") as handle:
    pickle.dump(df_str_text_rq2_ts_test, handle, protocol = pickle.HIGHEST_PROTOCOL)

    
with open("../Data for Model/X_train_rq2_ss_all.pkl", "wb") as handle:
    pickle.dump(df_str_text_rq2_ss_train, handle, protocol = pickle.HIGHEST_PROTOCOL)
    
with open("../Data for Model/X_test_rq2_ss_all.pkl", "wb") as handle:
    pickle.dump(df_str_text_rq2_ss_test, handle, protocol = pickle.HIGHEST_PROTOCOL)



### Change data to csvs for inspection
import glob
import os
import pickle
import pandas as pd
import re

filename_list = [file for file in glob.glob("../Data for Model/*.pkl")]

file_dict = {}
for file in filename_list:
    try:
        filename = open(file, "rb")
        f = pickle.load(filename)
        file_n = re.sub('.pkl', '', file)
        print(file_n)
        f.to_csv('{}.csv'.format(file_n))
    except(EOFError):
        break

### Checks to make sure indices are the same


os.chdir("C:\\Users\\[username]\\Downloads\\Updated Text Data") # insert [username]

filename_list = [file for file in glob.glob('../Data for Model/*test_rq2_ts_*.csv')]

for file in filename_list:
    df = pd.read_csv(file, index_col = 0)
    print(file)
    print(df.shape)
    print(df.index)

# run for train / test and rq1 / rq2 and ts / ss

os.chdir("C:\\Users\\[username]\\Downloads\\Updated Text Data") # insert [username]

sib = pd.read_csv('../Data for Model\siblings_train_rq1_ts_str.csv', index_col = 0, header = None)
X_test_all = pd.read_csv('../Data for Model\X_train_rq1_ts_all.csv', index_col = 0)
X_test_str = pd.read_csv('../Data for Model\X_train_rq1_ts_str.csv', index_col = 0)
y_test = pd.read_csv('../Data for Model\y_train_rq1_ts_str.csv', index_col = 0, header = None)

print(sib.shape)
print(X_test_all.shape)
print(X_test_str.shape)
print(y_test.shape)

print(sib.index)
print(X_test_all.index)
print(X_test_str.index)
print(y_test.index)

assert (X_test_all['PSID'].reset_index() == X_test_str['PSID'].reset_index()).all()['PSID']
assert (X_test_all['ReferralDatetime'].reset_index() == X_test_str['ReferralDatetime'].reset_index()).all()['ReferralDatetime']

X_test_all.set_index(X_test_str.index, inplace = True)
print(X_test_all.index)
X_test_all.to_csv('../Data for Model\X_train_rq1_ts_str_new.csv')

# Making sure indices match between siblings and y and X
import pandas as pd
import pickle
import os

os.chdir("C:\\Users\\[username]\\Downloads\\Updated Text Data") # insert [username]

sib = pd.read_csv('../Data for Model\siblings_train_rq2_ts_str.csv', index_col = 0, header = None)
X_train_all = pd.read_csv('../Data for Model\X_train_rq2_ts_all.csv', index_col = 0)
X_train_str = pd.read_csv('../Data for Model\X_train_rq2_ts_str.csv', index_col = 0)
y_train = pd.read_csv('../Data for Model\y_train_rq2_ts_str.csv', index_col = 0, header = None)
print(sib.shape)
print(X_train_all.shape)
print(X_train_str.shape)
print(y_train.shape)

# y alignment
# Take the index from the structured data
# Merge on PSID and ReferralDatetime
# Then reset the index from the structured data as for the all data
X_train_str['index'] = X_train_str.index
X_train_all_new = pd.merge(X_train_str[['PSID', 'ReferralDatetime', 'index']], X_train_all, on = ['PSID', 'ReferralDatetime'], how = 'left')
print(X_train_all_new.shape)
X_train_all_new.drop_duplicates(subset = ['index'], keep = 'first', inplace = True)
print(X_train_all_new.shape)
print(X_train_all_new.index)
print(X_train_all.index)
print(X_train_str.index)
X_train_all_new.set_index('index', drop = True, inplace = True)
print(X_train_all_new.index)

X_train_all_new.to_csv('../Data for Model\X_train_rq2_ts_str_new.csv')

### Siblings alignment
# Set working directory
os.chdir("C:\\Users\\[username]\\Downloads\\Updated Structured Data\\Created") # insert [username]

filename = open("df_outcome1_before_splitting.pkl", "rb")
df_outcome1_before_splitting = pickle.load(filename)
print(df_outcome1_before_splitting.shape)

filename = open("df_outcome2_before_splitting.pkl", "rb")
df_outcome2_before_splitting = pickle.load(filename)
print(df_outcome2_before_splitting.shape)

sib_new = pd.merge(X_train_str[['PSID', 'index']], df_outcome2_before_splitting[['PSID', 'PseudoID']], on = ['PSID'], how = 'left')

sib_new.drop_duplicates(subset = ['index'], keep = 'first', inplace = True)
print(sib_new.shape)
print(sib_new.index)
print(sib.index)
sib_new.set_index('index', drop = True, inplace = True)
print(sib_new.index)
sib_new.drop(columns = ['PSID'], inplace = True)

# Checking no siblings overlap
sib_test = pd.read_csv('../../Data for Model\siblings_test_rq2_ts_str.csv', index_col = 0, header = None)
assert len(set(sib_new['PseudoID']).intersection(set(sib_test[1]))) <= 1

sib_new.to_csv('../../Data for Model\siblings_train_rq2_ts_str_new.csv')

print(X_train_str.index)
print(X_train_all_new.index)
print(sib_new.index)
assert (X_train_str.index == X_train_all_new.index).all()
assert (X_train_str.index == sib_new.index).all()
assert (X_train_all_new.index == sib_new.index).all()

# Making sure indices match between siblings and y and X
import pandas as pd
import pickle
import os

os.chdir("C:\\Users\\[username]\\Downloads\\Updated Text Data") # insert [username]

sib = pd.read_csv('../Data for Model\siblings_test_rq2_ts_str.csv', index_col = 0, header = None)
X_test_all = pd.read_csv('../Data for Model\X_test_rq2_ts_all.csv', index_col = 0)
X_test_str = pd.read_csv('../Data for Model\X_test_rq2_ts_str.csv', index_col = 0)
y_test = pd.read_csv('../Data for Model\y_test_rq2_ts_str.csv', index_col = 0, header = None)
print(sib.shape)
print(X_test_all.shape)
print(X_test_str.shape)
print(y_test.shape)


print(sib.index)
print(X_test_all.index)
print(X_test_str.index)
print(y_test.index)

# y alignment
# Take the index from the structured data
# Merge on PSID and ReferralDatetime
# Then reset the index from the structured data as for the all data
X_test_str['index'] = X_test_str.index
X_test_all_new = pd.merge(X_test_str[['PSID', 'ReferralDatetime', 'index']], X_test_all, on = ['PSID', 'ReferralDatetime'], how = 'left')
print(X_test_all_new.shape)
X_test_all_new.drop_duplicates(subset = ['index'], keep = 'first', inplace = True)
print(X_test_all_new.shape)
print(X_test_all_new.index)
print(X_test_all.index)
print(X_test_str.index)
X_test_all_new.set_index('index', drop = True, inplace = True)
print(X_test_all_new.index)

X_test_all_new.to_csv('../Data for Model\X_test_rq2_ts_str_new.csv')

### Siblings alignment
# Set working directory
os.chdir("C:\\Users\\[username]\\Downloads\\Updated Structured Data\\Created") # insert [username]

filename = open("df_outcome1_before_splitting.pkl", "rb")
df_outcome1_before_splitting = pickle.load(filename)
print(df_outcome1_before_splitting.shape)

filename = open("df_outcome2_before_splitting.pkl", "rb")
df_outcome2_before_splitting = pickle.load(filename)
print(df_outcome2_before_splitting.shape)

sib_new = pd.merge(X_test_str[['PSID', 'index']], df_outcome2_before_splitting[['PSID', 'PseudoID']], on = ['PSID'], how = 'left')

sib_new.drop_duplicates(subset = ['index'], keep = 'first', inplace = True)
print(sib_new.shape)
print(sib_new.index)
print(sib.index)
sib_new.set_index('index', drop = True, inplace = True)
print(sib_new.index)
sib_new.drop(columns = ['PSID'], inplace = True)

# Checking no siblings overlap
sib_train = pd.read_csv('../../Data for Model\siblings_train_rq2_ts_str.csv', index_col = 0, header = None)
assert len(set(sib_new['PseudoID']).intersection(set(sib_train[1]))) <= 1

sib_new.to_csv('../../Data for Model\siblings_test_rq2_ts_str_new.csv')

print(X_test_str.index)
print(X_test_all_new.index)
print(sib_new.index)
assert (X_test_str.index == X_test_all_new.index).all()
assert (X_test_str.index == sib_new.index).all()
assert (X_test_all_new.index == sib_new.index).all()

# Making sure indices match between siblings and y and X
os.chdir("C:\\Users\\[username]\\Downloads\\Updated Text Data")

sib = pd.read_csv('../Data for Model\siblings_train_rq2_ss_str.csv', index_col = 0, header = None)
X_train_all = pd.read_csv('../Data for Model\X_train_rq2_ss_all.csv', index_col = 0)
X_train_str = pd.read_csv('../Data for Model\X_train_rq2_ss_str.csv', index_col = 0)
y_train = pd.read_csv('../Data for Model\y_train_rq2_ss_str.csv', index_col = 0, header = None)

print(sib.shape)
print(X_train_all.shape)
print(X_train_str.shape)
print(y_train.shape)

# y alignment
# Take the index from the structured data
# Merge on PSID and ReferralDatetime
# Then reset the index from the structured data as for the all data
X_train_str['index'] = X_train_str.index
X_train_all_new = pd.merge(X_train_str[['PSID', 'ReferralDatetime', 'index']], X_train_all, on = ['PSID', 'ReferralDatetime'], how = 'left')
print(X_train_all_new.shape)
X_train_all_new.drop_duplicates(subset = ['index'], keep = 'first', inplace = True)
print(X_train_all_new.shape)
print(X_train_all_new.index)
print(X_train_all.index)
print(X_train_str.index)
X_train_all_new.set_index('index', drop = True, inplace = True)
print(X_train_all_new.index)

X_train_all_new.to_csv('../Data for Model\X_train_rq2_ss_str_new.csv')

### Siblings alignment
# Set working directory
os.chdir("C:\\Users\\[username]\\Downloads\\Updated Structured Data\\Created")

filename = open("df_outcome1_before_splitting.pkl", "rb")
df_outcome1_before_splitting = pickle.load(filename)
print(df_outcome1_before_splitting.shape)

filename = open("df_outcome2_before_splitting.pkl", "rb")
df_outcome2_before_splitting = pickle.load(filename)
print(df_outcome2_before_splitting.shape)

sib_new = pd.merge(X_train_str[['PSID', 'index']], df_outcome2_before_splitting[['PSID', 'PseudoID']], on = ['PSID'], how = 'left')

sib_new.drop_duplicates(subset = ['index'], keep = 'first', inplace = True)
print(sib_new.shape)
print(sib_new.index)
print(sib.index)
print(X_train_str.index)
sib_new.set_index('index', drop = True, inplace = True)
print(sib_new.index)
sib_new.drop(columns = ['PSID'], inplace = True)

# Checking no siblings overlap
sib_test = pd.read_csv('../../Data for Model\siblings_test_rq2_ss_str.csv', index_col = 0, header = None)
assert len(set(sib_new['PseudoID']).intersection(set(sib_test[1]))) <= 1

sib_new.to_csv('../../Data for Model\siblings_train_rq2_ss_str_new.csv')

print(X_train_str.index)
print(X_train_all_new.index)
print(sib_new.index)
assert (X_train_str.index == X_train_all_new.index).all()
assert (X_train_str.index == sib_new.index).all()
assert (X_train_all_new.index == sib_new.index).all()

# Making sure indices match between siblings and y and X
os.chdir("C:\\Users\\[username]\\Downloads\\Updated Text Data")

sib = pd.read_csv('../Data for Model\siblings_test_rq2_ss_str.csv', index_col = 0, header = None)
X_test_all = pd.read_csv('../Data for Model\X_test_rq2_ss_all.csv', index_col = 0)
X_test_str = pd.read_csv('../Data for Model\X_test_rq2_ss_str.csv', index_col = 0)
y_test = pd.read_csv('../Data for Model\y_test_rq2_ss_str.csv', index_col = 0, header = None)

print(sib.shape)
print(X_test_all.shape)
print(X_test_str.shape)
print(y_test.shape)

print(sib.index)
print(X_test_all.index)
print(X_test_str.index)
print(y_test.index)

# y alignment
# Take the index from the structured data
# Merge on PSID and ReferralDatetime
# Then reset the index from the structured data as for the all data
X_test_str['index'] = X_test_str.index
X_test_all_new = pd.merge(X_test_str[['PSID', 'ReferralDatetime', 'index']], X_test_all, on = ['PSID', 'ReferralDatetime'], how = 'left')
print(X_test_all_new.shape)
X_test_all_new.drop_duplicates(subset = ['index'], keep = 'first', inplace = True)
print(X_test_all_new.shape)
print(X_test_all_new.index)
print(X_test_all.index)
print(X_test_str.index)
X_test_all_new.set_index('index', drop = True, inplace = True)
print(X_test_all_new.index)

X_test_all_new.to_csv('../Data for Model\X_test_rq2_ss_str_new.csv')

### Siblings alignment
# Set working directory
os.chdir("C:\\Users\\[username]\\Downloads\\Updated Structured Data\\Created")

filename = open("df_outcome1_before_splitting.pkl", "rb")
df_outcome1_before_splitting = pickle.load(filename)
print(df_outcome1_before_splitting.shape)

filename = open("df_outcome2_before_splitting.pkl", "rb")
df_outcome2_before_splitting = pickle.load(filename)
print(df_outcome2_before_splitting.shape)

sib_new = pd.merge(X_test_str[['PSID', 'index']], df_outcome2_before_splitting[['PSID', 'PseudoID']], on = ['PSID'], how = 'left')

sib_new.drop_duplicates(subset = ['index'], keep = 'first', inplace = True)
print(sib_new.shape)
print(sib_new.index)
print(sib.index)
print(X_test_str.index)
sib_new.set_index('index', drop = True, inplace = True)
print(sib_new.index)
sib_new.drop(columns = ['PSID'], inplace = True)

# Checking no siblings overlap
sib_train = pd.read_csv('../../Data for Model\siblings_train_rq2_ss_str.csv', index_col = 0, header = None)
assert len(set(sib_new['PseudoID']).intersection(set(sib_train[1]))) <= 1

sib_new.to_csv('../../Data for Model\siblings_test_rq2_ss_str_new.csv')

print(X_test_str.index)
print(X_test_all_new.index)
print(sib_new.index)
assert (X_test_str.index == X_test_all_new.index).all()
assert (X_test_str.index == sib_new.index).all()
assert (X_test_all_new.index == sib_new.index).all()

# Check new siblings

os.chdir("C:\\Users\\[username]\\Downloads\\")

sib_train = pd.read_csv('Data for Model\siblings_test_rq2_ts_str_new.csv', index_col = 0, header = None)
sib_train = pd.read_csv('Data for Model\siblings_train_rq2_ts_str_new.csv', index_col = 0, header = None)
assert len(set(sib_test[1]).intersection(set(sib_train[1]))) <= 1
