==> 7a_Handling_sensitive_data.ipynb <==
# Variable set-up

rq = 'rq1' # Options: 'rq1', 'rq2'
cv = 'ss' # Options: 'ts' (time series split, ignore siblings), 'ss' (stratified shuffle, ignore siblings)
data_type = 'all' # Options: 'str' (just structured data), 'all' (structured data and list of strings)

# Don't change
file_stub_y_siblings = rq + '_' + cv + '_str' # use 'str' for all 
file_stub = rq + '_' + cv + '_' + data_type # Creates file stub for saving in the format e.g. rq1_ss_str


## File directories
hard_drive_dir = '/Volumes/diskAshur2/Final transfer out data and code to WWC Jan 2020/Data for model/Use'


# Set working directory
import os
import pickle
os.chdir(hard_drive_dir)
os.getcwd()

# Import structured data -train, test data
import pandas as pd
X_tr = pd.read_csv("X_train_{}.csv".format(file_stub), index_col = 0)
print(X_tr.shape)
X_tr.reset_index(inplace = True, drop = True)
print(X_tr.index)

X_test = pd.read_csv("X_test_{}.csv".format(file_stub), index_col = 0)
print(X_test.shape)
X_test.reset_index(inplace = True, drop = True)
print(X_test.index)


import pandas as pd
y_tr = pd.read_csv("y_train_{}.csv".format(file_stub_y_siblings), index_col = 0, header = None)
print(y_tr.shape)
y_tr.reset_index(inplace = True, drop = True)
print(y_tr.index)

y_test = pd.read_csv("y_test_{}.csv".format(file_stub_y_siblings), index_col = 0, header = None)
print(y_test.shape)
y_test.reset_index(inplace = True, drop = True)
print(y_test.index)

assert X_tr.shape[0] == y_tr.shape[0]
assert X_test.shape[0] == y_test.shape[0]

print(X_tr.shape)
print(y_tr.shape)

print(set(X_test.columns).difference(set(X_tr.columns)))
print(set(X_tr.columns).difference(set(X_test.columns)))

# Sensitive data has the key columns
# And sensitive columns
X_tr_sensitive = X_tr[['PSID', 'ReferralDatetime', 'Gender',
       'AgeAtReferralDate', 'hasdisability', 'ethnicity_highlevel']]
print(X_tr_sensitive.columns)

X_test_sensitive = X_test[['PSID', 'ReferralDatetime', 'Gender',
       'AgeAtReferralDate', 'hasdisability', 'ethnicity_highlevel']]
print(X_test_sensitive.columns)

print(X_tr.select_dtypes(exclude = 'number').columns)
X_tr.drop(columns = ['detailed_ethnicity', 'ethnicity_highlevel'], inplace = True, errors = 'ignore')
print(X_tr.select_dtypes(exclude = 'number').columns)


print(X_test.select_dtypes(exclude = 'number').columns)
X_test.drop(columns = ['detailed_ethnicity', 'ethnicity_highlevel'], inplace = True, errors = 'ignore')
print(X_test.select_dtypes(exclude = 'number').columns)

print(X_tr.shape)
X_tr_gender = pd.get_dummies(X_tr['Gender'], prefix = 'Gender', dummy_na = False)
X_tr = pd.concat([X_tr, X_tr_gender], axis = 1)
X_tr.drop(columns = 'Gender', inplace = True)
print(X_tr.shape)

print(X_tr['hasdisability'].value_counts())
X_tr['hasdisability'].replace({True: 1, False: 0}, inplace = True)
print(X_tr['hasdisability'].value_counts())

print(X_tr['ReReferral'].value_counts())
X_tr['ReReferral'].replace({"Yes": 1, "No": 0}, inplace = True)
print(X_tr['ReReferral'].value_counts())

print(X_test.shape)
X_test_gender = pd.get_dummies(X_test['Gender'], prefix = 'Gender', dummy_na = False)
X_test = pd.concat([X_test, X_test_gender], axis = 1)
X_test.drop(columns = 'Gender', inplace = True)
print(X_test.shape)

print(X_test['hasdisability'].value_counts())
X_test['hasdisability'].replace({True: 1, False: 0}, inplace = True)
print(X_test['hasdisability'].value_counts())

print(X_test['ReReferral'].value_counts())
X_test['ReReferral'].replace({"Yes": 1, "No": 0}, inplace = True)
print(X_test['ReReferral'].value_counts())

print(X_tr.select_dtypes(exclude = 'number').columns)
print(X_test.select_dtypes(exclude = 'number').columns)

X_tr_sensitive.head()

X_test_sensitive.head()

assert 1==2

X_tr_sensitive.to_csv("X_train_sensitive_{}.csv".format(file_stub))
X_test_sensitive.to_csv("X_test_sensitive_{}.csv".format(file_stub))

X_tr.to_csv("X_train_{}.csv".format(file_stub))
X_test.to_csv("X_test_{}.csv".format(file_stub))

assert (set(X_test.columns).difference(set(X_tr.columns)) == {})
assert (set(X_tr.columns).difference(set(X_test.columns)) == {})

print(set(X_test.columns).difference(set(X_tr.columns)))
print(set(X_tr.columns).difference(set(X_test.columns)))

X_tr.to_csv("X_train_{}.csv".format(file_stub))
